Refactor Songsheet view
