+++ client/src/composables/index.js
@@ -5,3 +5,4 @@ export { default as useAutoScroll } from './useAutoScroll'
 export { default as useResponsive } from './useResponsive'
 export { default as useHideOnScroll } from './useHideOnScroll'
 export { default as useIonScroll } from './useIonScroll'
+export { default as useIonPageLifecycle } from './useIonPageLifecycle'
