+++ client/src/composables/useSongsheetParser.js
@@ -1,7 +1,7 @@
-import { ref, computed, watchEffect, toValue } from 'vue'
+import { ref, computed, watchEffect, toValue, reactive } from 'vue'
 import detectFormat from '@/lib/detect_format'
 import { Chord, ChordLyricsPair } from 'chordsheetjs'
-import useChords from '@/composables/useChords'
+import { useChords } from '@/composables'
 
 export default function useSongsheetParser (source) {
   const parser = computed(() => detectFormat(toValue(source)))
@@ -9,22 +9,30 @@ export default function useSongsheetParser (source) {
   const song = ref()
   const chords = useChords(song)
   const key = computed(() => song.value?.key ?? chords.value[0]?.toString())
+  const transpose = ref(0)
+  const capo = ref(0)
+  const transposedSong = computed(() => normalize(song.value?.transpose(transpose.value - capo.value)))
+  const transposed = reactive({
+    key: computed(() => Chord.parse(key.value).transpose(transpose.value, { normalizeChordSuffix: true })),
+    song: transposedSong,
+    chords: useChords(transposedSong)
+  })
 
-  // Using watchEffect instead of computed to try to parse the source whenever it changes
-  // so we can keep the previously parsed song if it fails.
+  // Using watchEffect instead of computed so we can keep the previously parsed song if it fails.
   watchEffect(() => {
     try {
-      song.value = normalize(parser.value?.parse(toValue(source)))
-
-      // parsing succeeded, so clear last error
-      error.value = null
+      const parsed = parser.value?.parse(toValue(source))
+      const originalCapo = Number(parsed.capo ?? 0)
+      capo.value = originalCapo
+      song.value = normalize(parsed.transpose(originalCapo, { normalizeChordSuffix: true }))
+      error.value = null // parsing succeeded, so clear last error
     } catch (e) {
       error.value = e
       return null
     }
   })
 
-  return { song, key, chords, error, parser }
+  return { song, key, chords, error, parser, transpose, capo, transposed }
 }
 
 function normalize (song) {
